Initialize a themed shiny app

We are working to make high level function which will initialize a boilerplate shiny application with certain features already included; for now these include branding and a loading page. In the future we will work to add additional modules as specified by the user. For example an the addition of boilerplate code for sync-maps, sankeys ... any other type of visualization that we have previous already developed. This hopefully will allow open-sourcing of high level shiny development and making getting started with Shiny easier. Right now two modules are loaded: 1) a generic ggplot histogram 2) a sync-map component.

Usage

initializeShinyApp(dest = "./", theme = "SALURBAL")

Arguments

dest

the path to initialize your app. Important: you should not specify the file name, just the directory. './folder/app.R' is wrong. './folder/' is correct. A folder will be created with an app.R file already initialized!

theme

which theme do you want: Two options are available "UHC" or "SALURBAL"

Value

A character vector.

Examples

initializeShinyApp(dest = "./demo", theme = "SALURBAL")
#> Error in initializeShinyApp(dest = "./demo", theme = "SALURBAL"): Output directory already exists, please choose a new directory.